requirements.txt
alembic==1.14.1
annotated-types==0.7.0
anyio==4.8.0
asyncpg==0.30.0
black==24.10.0
certifi==2024.12.14
cfgv==3.4.0
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
distlib==0.3.9
fastapi==0.115.7
filelock==3.17.0
flake8==7.1.1
greenlet==3.1.1
h11==0.14.0
identify==2.6.6
idna==3.10
Mako==1.3.8
MarkupSafe==3.0.2
mccabe==0.7.0
mypy-extensions==1.0.0
nodeenv==1.9.1
packaging==24.2
pathspec==0.12.1
platformdirs==4.3.6
pre_commit==4.1.0
psycopg2==2.9.10
psycopg2-binary==2.9.10
pycodestyle==2.12.1
pydantic==2.10.6
pydantic-settings==2.7.1
pydantic_core==2.27.2
pyflakes==3.2.0
python-dotenv==1.0.1
PyYAML==6.0.2
requests==2.32.3
sniffio==1.3.1
SQLAlchemy==2.0.37
starlette==0.45.3
typing_extensions==4.12.2
urllib3==2.3.0
uvicorn==0.34.0
virtualenv==20.29.1
